Reorganize/refactor packet definitions and add missing doc comments
