+++ Program/PacketParsing/Packets/Guitar/GuitarInput.cs
