+++ Program/RB4InstrumentMapper.csproj
@@ -129,16 +129,16 @@
     <Compile Include="PacketParsing\Mappers\MapperFactory.cs" />
     <Compile Include="PacketParsing\Mappers\VigemMapper.cs" />
     <Compile Include="PacketParsing\Mappers\VjoyMapper.cs" />
-    <Compile Include="PacketParsing\Packets\Acknowledgement.cs" />
+    <Compile Include="PacketParsing\Packets\Drums\DrumInput.cs" />
+    <Compile Include="PacketParsing\Packets\Gamepad\GamepadInput.cs" />
+    <Compile Include="PacketParsing\Packets\Guitar\GuitarInput.cs" />
+    <Compile Include="PacketParsing\Packets\System\Acknowledgement.cs" />
+    <Compile Include="PacketParsing\Packets\System\DeviceArrival.cs" />
+    <Compile Include="PacketParsing\Packets\System\DeviceStatus.cs" />
+    <Compile Include="PacketParsing\Packets\System\Keystroke.cs" />
+    <Compile Include="PacketParsing\Packets\System\XboxDescriptor.cs" />
     <Compile Include="PacketParsing\Packets\ChunkBuffer.cs" />
     <Compile Include="PacketParsing\Packets\CommandHeader.cs" />
-    <Compile Include="PacketParsing\Packets\DeviceArrival.cs" />
-    <Compile Include="PacketParsing\Packets\DeviceStatus.cs" />
-    <Compile Include="PacketParsing\Packets\DrumInput.cs" />
-    <Compile Include="PacketParsing\Packets\GamepadInput.cs" />
-    <Compile Include="PacketParsing\Packets\GuitarInput.cs" />
-    <Compile Include="PacketParsing\Packets\Keystroke.cs" />
-    <Compile Include="PacketParsing\Packets\XboxDescriptor.cs" />
     <Compile Include="PacketParsing\Packets\XboxMessage.cs" />
     <Compile Include="PacketParsing\DeviceGuids.cs" />
     <Compile Include="PacketParsing\ParsingUtils.cs" />
